+++ src/django_mysql/models/fields/lists.py
@@ -1,4 +1,3 @@
-import django
 from django.core import checks
 from django.db.models import CharField, IntegerField, Lookup, TextField
 from django.utils.translation import gettext_lazy as _
@@ -87,27 +86,13 @@ class ListFieldMixin:
                 value = [self.base_field.to_python(v) for v in value.split(",")]
         return value
 
-    if django.VERSION >= (2, 0):
-
-        def from_db_value(self, value, expression, connection):
-            # Similar to to_python, for Django 1.8+
-            if isinstance(value, str):
-                if not len(value):
-                    value = []
-                else:
-                    value = [self.base_field.to_python(v) for v in value.split(",")]
-            return value
-
-    else:
-
-        def from_db_value(self, value, expression, connection, context):
-            # Similar to to_python, for Django 1.8+
-            if isinstance(value, str):
-                if not len(value):
-                    value = []
-                else:
-                    value = [self.base_field.to_python(v) for v in value.split(",")]
-            return value
+    def from_db_value(self, value, expression, connection):
+        if isinstance(value, str):
+            if not len(value):
+                value = []
+            else:
+                value = [self.base_field.to_python(v) for v in value.split(",")]
+        return value
 
     def get_prep_value(self, value):
         if isinstance(value, list):