+++ HISTORY.rst
@@ -7,6 +7,7 @@ Pending
 -------
 
 * Prevent ``collections.abc.Sequence`` warning.
+* Drop Django 1.11 support. Only Django 2.0+ is supported now.
 
 3.3.0 (2019-12-10)
 ------------------