+++ HISTORY.rst
@@ -8,6 +8,8 @@ Pending
 
 * Prevent ``collections.abc.Sequence`` warning.
 * Drop Django 1.11 support. Only Django 2.0+ is supported now.
+* Prevent ``JSONField`` from adding ``CAST(... AS JSON)`` to queries against
+  ``str``, ``int``, and ``float`` types.
 
 3.3.0 (2019-12-10)
 ------------------